renamed job name from samplejob to sample in step 4
While creating a new scheduler in Azure portal, scheduler name must
match the name of the job created in solution EXCLUDING the 'Job'
keyword. For example for creating a scheduled job for SampleJob, it must
be named as "Sample". If SampleJob is used, it will result in an 404
error.

In this commit:
-Fixed the 2 images which were showing SampleJob to Sample
-Fixed the article instruction 4 to name scheduler as Sample in place of
SampleJob