Update mobile-engagement-user-interface.md
-	New Applications (Including the log screenshot, the + button in another screenshot)
o	I removed the paragraphs that referred to creating new applications from within the portal (I also removed the image of iOS, Android, and Windows logos).
o	I was unable to upload the new copy of the image that does not include the “New Applications” button, but it the edited image attached to my e-mail in case anyone can update the file into the folder “media/ mobile-engagement-user-interface/home1.png” for me. Thanks.
-	Mention of Demo Applications
o	I removed all references to the demo applications.
-	Favorite applications (it is there now but will be removed in a week)
o	I removed the paragraphs that referred to managing favorites (including the link to the image).