Updated table columns.
It seems Github may have implemented a stricter interpretation of markdown. In markdown pad, the code works when the markdown flavor is set to "extended" and git-hub (offline) and I would have noticed a broken table during my review. Adding an extra column to the dividing line seems to have corrected the problem.