Update xplat-cli-install.md
It was confusing how to actually run the 'azure' command.  it took quite a while to figure out you must have an Administrator cmd.exe for it to work, otherwise you get "'azure' not recognized". In addition the note about "you can run it anywhere" felt like it was tacked onto the npm/node.js section.  This change pulls out that note into an explicit section to make it clear and puts the example for Windows, with admin privs, there plus the command line to make it super clear.  Should probably look at adding more examples for other environments as well.