Update documentdb-get-started.md
Added explicit instructions on where some code needed to be placed in the file. Also changed the Step 1 #2 from "Select New from the File menu" to "On the **File** menu, select **New**" as you want to present UI items in the order in which they appear on the screen. I ran the tutorial and it works as expected. The keys.png screenshot is out of date - I will update it as it's the same screenshot as is used in many other files.