Update sql-database-get-started-sql-data-sync.md
Clearing up confusion when the tutorial starts by saying it is using the "Preview" portal but actually is using the older Management portal. Back when SQL Azure Data Sync was first available the HTML 5 based management portal was the Preview portal, but now it's the old portal and Ibiza is the preview portal.