Update event-hubs-overview.md
There are two terms used intermittently: "event publishers" and "event producers". They seem to mean the exact same thing. However, since "event producers" is used only a few times in the document, this confused me. 

It is my suggestion to replace the three instances of "event producers" with "event publishers", as is consistent with the diagrams.