Update app-insights-windows-desktop.md
The github link
    https://github.com/Microsoft/ApplicationInsights-dotnet/tree/master/src/TelemetryChannels/PersistenceChannel

at the bottom of this page is broken.

Reason:
Persistence Channel has been renamed to ServerPersistenceChannel and as its name suggests is only used for server side. 
v1.2.3 is the latest NuGet/Github package that is used for windows.  

Suggestion:
Change both links, the NuGet and GitHub, to point to v1.2.3 
NuGet:
   https://www.nuget.org/packages/Microsoft.ApplicationInsights.PersistenceChannel/1.2.3

GitHub:
   https://github.com/Microsoft/ApplicationInsights-dotnet/tree/v1.2.3/src/TelemetryChannels/PersistenceChannel