Update azure-toolkit-for-eclipse-whats-new.md
[Context]
This helps enable remote debugging and JMX diagnostics for specific compute instances running in the cloud in scenarios with multi -instance deployments.{j}(Instance endpoints are also listed in
[Source]
(Instance endpoints are also listed in
[Query]
The part "(Instance endpoints are also listed in" is incomplete. It seems to be a source error.