Update guidance-multitenant-identity-adfs.md
typo correction based on the screen shot 
Typo. "Pass Through of Filter an Incoming Claim" should be "Pass Through or Filter an Incoming Claim." 
I checked the screen shot on the page of this article. And "Pass Through or Filter an Incoming Claim" is displayed there, 
not "Pass Through of Filter an Incoming Claim."