Fixing some typos
It’s been almost a year since we announced the support for Virtual Machines under Azure Resource Manager. You can read more about the advancements and the additional capabilities that it supports >>>here<<< no link.
And there’s a focus of enabling security by default with the enforcement of virtual machines in a Virtual Network. - shouldn't it be "focus ON"?

I corrected a few is\are here and there, not really sure i'm 100% correct, please check. I'm not a native English speaker. Thanks!