Updated to use SAS in place of ACS
ACS is no longer supported in Azure Service Bus Queues, so we now need to leverage the SAS policy name and key in place of the UserName and Password values. I've tested this change for a client that was trying to do JMS to ASB over AMQP and it worked. Also, note that this feedback was previously provided in the comments section for this article.