Fixed Typo in EnableBGP Flag
In the article the EnableBGP Flag uses the parameter "True", although PowerShell will accept this and create a connection, the connections status will remain unknown when doing a cross subscription VNET-to-VNET BGP Connection. The flag must be set $True when creating the connection for the connection status to be "Connected" and allow traffic between the subscriptions.