Fixed typo in step to set context and restore JSON file
There were a couple typos in the steps where you set the storage context and restore the JSON config file. I also don't think the -context parameter was needed on the Get-AzureStorageBlobContent command as we set the current context previously, and it doesn't look like that variable was actually ever set in the first place.