Updated link
Thanks for the suggestion.  There is no need to add "articles" to this link path because the article is in \articles, as-is the article it links to. In this case, only the article name being linked to is required, which is what the original link was.  The "/" that was between the "md" and "#" shouldn't have been there though, so I removed it.