Step 2 line 60 and Line 74
In swaggerize steps, docs say to enter project name "contactlist", all lower case. In the code screenshot, "ContactList" with caps rather than all lowercase is entered. In the next steps docs say to cd to "ContactList". The git clone has a "ContactList" folder so that is what you would be cd'ing to, as I did. I also changed line 74 to give the path to the subfolder, just to be clear as possible. Hope this helps.