M663084: Unclear sentences in line 41
@andkjell,
The sentences in line 41: "The most common change is changes to the attribute flows. The data in your source directory might not be how you want it in Azure AD." seem not to be clear enough. Please, check if the proposed sentence change meets your idea and sign off if you agree.
Many thanks in advance.
CC: @v-jacard