Clarify authentication certificate and probe relationship
Added additional note in Step 7 to clarify the usage of the authentication certificate and how it correlates with the default probe's behaviour (in lieu of working custom HTTPS probes) in the event that the back-ends host multiple sites with SSL bindings and/or SNI. Since the probe only reaches the default SSL binding; if there *is no* default SSL binding, the back-end would not be whitelisted. If there *is* a binding and the public key does not match the authentication certificate, the back-end would not be whitelisted. Therefore, the public key of the default SSL binding of the back-end(s) must match the provided authentication certificate. In many cases this would mean that the same authentication certificate is used for ALL host-header/SNI enabled SSL sites hosted on the same back-end.