The example is not valid
The example under ##Working with Date Times is not valid for the schema 2016-06-01. The timingWarning action should be in a separate "if" action. It looks like the author started to make that change, since the timingWarning action contains directly anotheraction of no type (and is invalid).

I finished it and made it valid.