Update web-sites-dotnet-deploy-aspnet-mvc-app-membership-oauth-sql-database.md
Step 3 of "Protect the Application with SSL and the Authorize Attribute", the ContactsController.cs file that was created at the start of the project bears the name "CmController.cs" here. This has been corrected to read "ContactsController.cs" to match the image when adding this controller at the beginning of this tutorial.