Update per feedback
Right now the metadata title is missing the word "Azure" in the first
part per our metadata guidelines.
https://github.com/Microsoft/azure-docs-pr/blob/master/contributor-guide/article-metadata.md
Also, there is an en-US locale encoded in the document. Per our guidance
"Links to pages on azure.microsoft.com, TechNet, and MSDN are coded as
locale agnostic. Do not include en-us, en-gb, en-in, or any other locale
in links to these sites."
Finally - the resolution for switch.jpg is low and the graphic is hardly
readable. Please improve the resolution on the graphic if you can.