Add example of using onboarding config
For both the manual and cmdlet methods of generating an onboarding config, I added a usage example of `Set-DscLocalConfigurationManager`. For someone unfamiliar with DSC, consuming a metaconfig could be a non-obvious step. 

I also added a link to the DSC documentation on metaconfigurations (which, weirdly enough, is called `metaconfig` in its friendly URL, but doesn't mention the term "metaconfig" once...that should probably be corrected at some point).