Update expressroute-troubleshooting-expressroute-overview.md
Proposing an update to the acronyms that are used in this document to line up with the image that shows how the connectivity works. Currently, the document references MSEE-PRs, which is not defined in the document though is noted as representing the PE equipment that faces the MSEE routers. MSEE-PR is somewhat confusing and ambiguous, so I'm proposing we more clearly define this by calling this PE-MSEE instead, representing that this is the provider edge equipment connecting to the MS Enterprise edge equipment. We could also put a definition around MSEE-PR and keep using it, but I think that changing this makes a little more sense and lines up better with what the visualization shows. Welcome any and all feedback.