Update the C# word count program
The C# word count program does not count word occurrences. Instead it counts words. This is not in sync with the description and the script earlier in the article that states that the program counts word count occurrences. My fix will make the job definition script pass and also will make the behavior and output match what is described in the article.