M1976: Possible incorrect format: "The Reject Type and Reject Value allow you to define how many rows or what percentage of the data must be present in the final table"
@cakarst @barbkess
It seems that it should say "Reject_Type" and "Reject_Value, instead of "Reject Type" and "Reject Value" to align with the code. The main verb should be written without the "s" at the end of the word since the subject is plural. 
Could you confirm this, please?
Many thanks in advance.