Update service-bus-nodejs-how-to-use-queues.md
This updates the doc to use the `AZURE_SERVICEBUS_CONNECTION_STRING` environment variable instead of the `AZURE_SERVICEBUS_NAMESPACE` and `AZURE_SERVICEBUS_ACCESS_KEY`. If you only specify those two envrinment variables, you get a DNS resolution error from inside the SDK. I _think_ you need to also specify `AZURE_SERVICEBUS_ISSUER` if you specify `AZURE_SERVICEBUS_NAMESPACE` and `AZURE_SERVICEBUS_ACCESS_KEY`, but I was not able to get it to work. This way is much simpler, and is verified to work as of this writing.