Correct rand doc: the second end isn't inclusive
Hi

It was said in the doc that both ends where inclusive in the rand function.
I wanted to have either '0' or '1'.
By testing the 'rand(0,1)' function as the doc suggested, in a 100 times loop, I got many times 100 times '0'. I didn't get any '1'.
By testing the 'rand(0,2)' function, I got both '0' and '1'.

So I changed the doc to reflect that. Feel free to give an other exemple if this one doesn't fit you :)

Julien