Corrected application-gateway create command
When running "az network application-gateway create", an error is returned when vnet-address-prefix and subnet-address-prefix are specified along with vnet-name and subnet specified, so removed the two *-prefix params. Also, this doc is listed under the "Public Facing" section but it doesn't actually create a public IP, so I added --public-ip-address.