Update hdinsight-domain-joined-architecture.md
Updated "Azure AD" to "AD" in several places in intro sections as well as section "HDInsight integrated with AD running on Azure IaaS". The 2 supported options now include 1 with AD and 1 with Azure AD. Therefore, the article intro section should use the term "AD" as a generic term encompassing both. And the section with AD (not AAD) had referred erroneously to "Azure AD", so needed to be updated to "AD".